Add prime number generation
The method used to create this was split into two parts.
the first part is checking whether a number is prime,
the second part is just enumurating through numbers until
we have filled an array with enough prime numbers, using the
aforementioned prime number check.

for the prime number generation, this method should be fast enough
to generate an output for a terminal appllication, so i optimised
for readability.

we kept the api the same as the ruby prime, so the first method will return
the first prime number, and if you ask for the first(5) it will give you
the first five in an array.

I am not a massive fan of the fact it can return  two different types,
but for the output required,if this causes problems, i can refactor it.
this seemed the most rubyesque way to do it.

as far the prime number generation, it basically just divides with all numbers
in the square root which is explained in links below [1][2].

[1]https://stackoverflow.com/a/5813926
[2]http://www.studyalgorithms.com/misc/find-the-first-n-prime-numbers-method-1/