Add table generations, and change api for .first when no argument is given
so this adds in the basic code for table generation, this just outputs
the basic multiplication table without anything extra. the table also
takes the input of the numbers, so it can be reused if needed.

We also changed the api to .first, as this would cause problems with
the table generation. This makes me more comfortable as .first now always
returns the same type. in a traditional world, this would probably be in
a whole seperate pull, but for the example, i think it is ok to change this here