Add the terminal table generation in dn first pass at output
this adds in the terminal table generation. This is responsible for actually
generating the terminal table output

I have also put in the first look at what the table actually looks like.

The table is currently missing the left hand side column with the number,
this will be added in in the next commit, ihave also hardcoded the number
in the terminal for the moment, as i will deal with that as well,
i need to check if the number is valid etc