Cleanup row generation with header and get terminal script complete
This adds in the the row starting headers of the prime number, and
also cleans up the terminal script.

i'm not 100% happy with how it gets the row header,as it is
a biot disconencted from the row, and expansion may cause
problems there. given more time i would of added this into
the row generation int he table, but i ran out of time to
think about that properly. also not 100% convinced about the
parsing of the bad arguments, but seems to work for now and
to show the general concept